refactor(uffd-tests): collapse cross-process harness side-channels into RPC
Reduce parent ↔ child side channels from {7 env config vars, 1 helper
flag, 1 uffd ExtraFile, 1 content tmp file, 1 RPC socket} to
{2 env vars, 1 uffd ExtraFile, 1 RPC socket}. Everything that was
configured via env or the content file now flows through a single
Lifecycle.Bootstrap RPC.

Drops:
- 7 config env vars (GO_UFFD_SOCKET kept; GO_UFFD_CONTENT, GO_UFFD_MMAP_*,
  GO_UFFD_ALWAYS_WP, GO_UFFD_GATED, GO_UFFD_BARRIERS removed)
- the content.bin tmp file (passed as []byte in BootstrapArgs)
- the testConfig.gated flag and its conditional RPC wiring
- childForkMu — replaced with a single F_DUPFD_CLOEXEC fcntl that
  removes the dup-then-set-cloexec race window the mutex was guarding

Splits the 752-line cross_process_helpers_test.go into seven cohesive
files: harness_{parent,child,client}_test.go for the wire/IPC plumbing
and rpc_{lifecycle,paging,barriers}_test.go for the three RPC services.
The Service god-struct becomes three small services multi-registered
with net/rpc; the parent calls Lifecycle.Bootstrap, Paging.Pause,
Barriers.Install, etc. via a typed *harnessClient wrapper.

testHandler now holds a single *harnessClient instead of
{client, conn, cmd} plus servePause/serveResume/pageStatesOnce
function fields; pageStatesOnce becomes a regular method that
delegates to the client.

Test-only; no production code changes. go vet / golangci-lint /
\`sudo go test -race\` all clean (×2).
